requirements.txt
alembic==1.3.3
asgiref==3.2.3
attrs==19.3.0
Automat==0.8.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.3
configparser==4.0.2
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
databricks-cli==0.9.1
decorator==4.4.1
defusedxml==0.6.0
Django==3.0.2
django-background-task==0.1.8
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.8.1
djangorestframework==3.11.0
docker==4.1.0
entrypoints==0.3
enum34==1.1.6
facenet-pytorch==2.2.9
Flask==1.1.1
gitdb2==2.0.6
GitPython==3.0.5
gorilla==0.3.0
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.4.0
incremental==17.5.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
json5==0.8.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyter-http-over-ws==0.0.7
jupyterlab==1.2.5
jupyterlab-server==1.0.6
kiwisolver==1.1.0
lxml==4.4.2
Mako==1.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==8.1.0
nbconvert==5.6.1
nbformat==5.0.4
notebook== 6.0.3
numpy==1.18.1
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
opencv-python-headless==4.2.0.34
pandas==0.25.3
pandocfilters==1.4.2
parsel==1.5.2
parso==0.5.2
pickleshare==0.7.5
Pillow==7.0.0
pipenv==2018.11.26
prometheus-client==0.7.1
prometheus-flask-exporter==0.12.1
prompt-toolkit==3.0.2
Protego==0.1.16
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.5.2
PyHamcrest==2.0.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pypiwin32==223
pyrsistent==0.15.7
python-dateutil==2.8.1
python-editor==1.0.4
python-scrapyd-api==2.1.2
pytz==2019.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.3
pyzmq==18.1.1
querystring-parser==1.2.4
queuelib==1.5.0
requests==2.22.0
scipy==1.4.1
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.17.0
six==1.14.0
smmap2==2.0.5
soupsieve==2.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
tabulate==0.8.6
terminado==0.8.3
testpath==0.4.4
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.5.0+cpu 
torchvision==0.6.0+cpu
tornado==6.0.3
traitlets==4.3.3
virtualenv==16.7.9
virtualenv-clone==0.5.3
w3lib==1.21.0
waitress==1.4.2
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.0
wrapt==1.11.2
zipp==2.0.1
zope.interface==4.7.1